Restructure to follow clean + extra cleanup
