+++ Application/DependencyInjection.cs
@@ -1,4 +1,5 @@
-﻿namespace Application;
+﻿
+namespace Application;
 
 public static class DependencyInjection
 {
@@ -7,7 +8,7 @@ public static class DependencyInjection
         var assembly = typeof(DependencyInjection).Assembly;
         services.AddValidatorsFromAssembly(assembly);
         services.AddScoped<IKinoJoinDbService, KinoJoinDbService>();
-        services.AddScoped<IFetchNewestKinoDkDataService, FetchNewestKinoDkDataService>();
+        // services.AddScoped<IFetchNewestKinoDkDataService, FetchNewestKinoDkDataService>();
         services.AddScoped<ICalendarService, CalendarService>();
         return services;
     }
