+++ Application/Interfaces/IKinoJoinDbService.cs
@@ -1,5 +1,4 @@
-﻿using System.Linq.Expressions;
-
+﻿
 namespace Application.Interfaces;
 
 /**
