+++ Application/Interfaces/IUserInfoService.cs
@@ -1,5 +1,4 @@
-﻿using Microsoft.AspNetCore.Components.Authorization;
-
+﻿
 namespace Application.Interfaces;
 
 /// <summary>
