+++ Domain/Usings.cs
@@ -1,5 +1,3 @@
 ﻿global using System.ComponentModel.DataAnnotations;
 global using System.ComponentModel.DataAnnotations.Schema;
 global using Microsoft.EntityFrameworkCore;
-global using Newtonsoft.Json;
-global using Newtonsoft.Json.Linq;
