+++ Infrastructure/Infrastructure.csproj
@@ -29,4 +29,8 @@
         <ProjectReference Include="..\Application\Application.csproj"/>
     </ItemGroup>
 
+    <ItemGroup>
+        <Folder Include="Interfaces\"/>
+    </ItemGroup>
+
 </Project>
