+++ Infrastructure/KinoAPI/FilterApiHandler.cs
@@ -1,4 +1,4 @@
-namespace Application.Services;
+namespace Infrastructure.KinoAPI;
 
 public class FilterApiHandler : IFilterApiHandler
 {
@@ -335,11 +335,4 @@ public class FilterApiHandler : IFilterApiHandler
 
         return (selectedCinemas, selectedMovies, selectedGenres, startDate, endDate, sortOption);
     }
-}
-
-public enum SortOption
-{
-    Most_viewed,
-    Alphabetical,
-    Rating,
 }
