+++ Presentation/Presentation.Client/Pages/Components/Shared/Showtimes/ShowtimeSelector.razor
@@ -1,4 +1,3 @@
-@using Application.Services
 @inject IJSRuntime JsRuntime
 
 @*
