+++ Presentation/Presentation.Client/Pages/JoinCreate.razor
@@ -1,6 +1,5 @@
 ﻿@page "/"
 @page "/create-event/{UrlFilterString?}"
-@using Application.Services
 @inject NavigationManager NavigationManager
 @inject IJSRuntime JsRuntime
 @inject KinoJoinHttpClient KinoJoinHttpClient
