+++ Test/KinoJoin/KinoJoinTests.cs
@@ -1,10 +1,10 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.Net;
 using System.Net.Http.Json;
-using Application.Services;
 using Domain.Entities;
 using Domain.ExternalApiModels;
 using FluentAssertions;
+using Infrastructure.KinoAPI;
 using Infrastructure.Persistence;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
